Standardized API POJOs, Cleaned Up, Ready for Calculating Player Statistics
