+++ src/main/java/com/wotos/wotosplayerservice/repo/PlayerStatisticsSnapshotsRepository.java
@@ -1,0 +1,7 @@
+package com.wotos.wotosplayerservice.repo;
+
+import com.wotos.wotosplayerservice.dao.PlayerStatisticsSnapshot;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PlayerStatisticsSnapshotsRepository extends JpaRepository<PlayerStatisticsSnapshot, Integer> {
+}
