+++ src/main/java/com/wotos/wotosplayerservice/util/feign/WotAccountsFeignClient.java
@@ -2,6 +2,10 @@ package com.wotos.wotosplayerservice.util.feign;
 
 import com.wotos.wotosplayerservice.config.FeignConfig;
 import com.wotos.wotosplayerservice.util.model.wot.*;
+import com.wotos.wotosplayerservice.util.model.wot.achievements.Achievements;
+import com.wotos.wotosplayerservice.util.model.wot.player.Player;
+import com.wotos.wotosplayerservice.util.model.wot.player.PlayerDetails;
+import com.wotos.wotosplayerservice.util.model.wot.player.PlayerVehicle;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -27,24 +31,27 @@ public interface WotAccountsFeignClient {
     @GetMapping(value = "/info/", consumes = "application/json")
     ResponseEntity<WotApiResponse<Map<String, PlayerDetails>>> getPlayerDetails(
             @RequestParam("application_id") String appId,
+            @RequestParam("access_token") String accessToken,
+            @RequestParam("extra") String extra,
+            @RequestParam("fields") String fields,
             @RequestParam("language") String language,
-            @RequestParam("account_id") String accountId
+            @RequestParam("account_id") Integer accountId
     );
 
     @GetMapping(value = "/tanks/")
     ResponseEntity<WotApiResponse<List<PlayerVehicle>>> getPlayerVehicles(
             @RequestParam("application_id") String appId,
-            @RequestParam("account_id") String accountId,
+            @RequestParam("account_id") Integer accountId,
             @RequestParam("access_token") String accessToken,
             @RequestParam("fields") String fields,
             @RequestParam("language") String language,
-            @RequestParam("tank_id") String tankIds
+            @RequestParam("tank_id") Integer[] tankIds
     );
 
     @GetMapping(value = "/achievements/")
     ResponseEntity<WotApiResponse<Achievements>> getPlayerAchievements(
             @RequestParam("application_id") String appId,
-            @RequestParam("account_id") String accountId,
+            @RequestParam("account_id") Integer accountId,
             @RequestParam("fields") String fields,
             @RequestParam("language") String language
     );
